87 sweet alert로 경고하기

+++ src/components/display-comment.vue
@@ -5,7 +5,7 @@
         <v-textarea
           v-model="comment"
           outlined
-          label="댓글 작성!"
+          label="댓글 작성"
           placeholder="Ctrl + Enter로 작성 가능"
           append-icon="mdi-comment-plus"
           @click:append="save"
@@ -298,6 +298,14 @@ export default {
       comment.likeUids = item.likeUids
     },
     async remove (comment) {
+      const r = await this.$swal.fire({
+        title: '정말 삭제하시겠습니까?',
+        text: '삭제 후 되돌릴 수 없습니다.',
+        icon: 'error',
+        // confirmButtonText: 'Cool',
+        showCancelButton: true
+      })
+      if (!r.value) return
       await this.docRef.collection('comments').doc(comment.id).delete()
       const i = this.items.findIndex(el => el.id === comment.id)
       this.items.splice(i, 1)
